In theory this should work, although copilot generated it all and it still needs to be reviewed.

+++ app/api/submitKill/route.ts
@@ -24,25 +24,38 @@ export async function POST(request: Request) {
     const data: SubmitKillPayload = await request.json();
     const prisma = new PrismaClient()
 
-    // TODO: Reimplement this
-    /*let lastKill;
+    let ffa;
     try {
-      lastKill = await prisma.kill.findFirst({
+      ffa = await prisma.gameConfiguration.findUnique({
         where: {
-          killerId: session.user.id
-        },
-        orderBy: {
-          createdAt: "asc"
+          key: "ffa"
         }
       })
     } catch {
-      return new NextResponse("Internal error (1)", { status: 500 })
+      return new NextResponse("Internal error (3)", { status: 500 })
     }
 
-    if(lastKill && (lastKill.createdAt.getTime()+1000*60) > Date.now()) return new NextResponse(
-      "You must wait 15 minutes since your last kill submission to submit another", {
-        status: 450
-      })*/
+    if(ffa) {
+      let lastKill;
+      try {
+        lastKill = await prisma.kill.findFirst({
+          where: {
+            killerId: session.user.id
+          },
+          orderBy: {
+            createdAt: "desc"
+          }
+        })
+      } catch {
+        return new NextResponse("Internal error (1)", { status: 500 })
+      }
+
+      // TODO: Change this to 15 minutes
+      if(lastKill && (lastKill.createdAt.getTime()+1000) > Date.now()) return new NextResponse(
+        "You must wait 15 minutes since your last kill submission to submit another", {
+          status: 450
+        })
+    }
 
     if(!session.user.currentTarget) return new NextResponse(
       "Error finding your target. Contact the spoonmaster to resolve. (1)", {
@@ -64,17 +77,6 @@ export async function POST(request: Request) {
       "Error finding your target. Contact the spoonmaster to resolve. (2)", {
         status: 500
       })
-
-    let ffa;
-    try {
-      ffa = await prisma.gameConfiguration.findUnique({
-        where: {
-          key: "ffa"
-        }
-      })
-    } catch {
-      return new NextResponse("Internal error (3)", { status: 500 })
-    }
 
     let error = false;
